Use index.tsx according to guidelines
